+++ c/tmsim/player.h
@@ -4,7 +4,6 @@
 /***************************************
 * TMSim - Player header file
 ***************************************/
-
 struct player {
     // player stats
     uint8_t intelligence;
@@ -21,10 +20,10 @@ struct player {
 };
 
 // rolls a single stat based on a 1-5 difficulty
-int roll_stat(int);
+uint8_t roll_stat(uint8_t);
 
 // roll a player's stats based on a 1-5 difficulty
-void roll_stats(int, struct player*);
+void roll_stats(uint8_t, struct player*);
 
 // get a player struct with initialized stats
-struct player get_player(int);
+struct player get_player(uint8_t);
